Return to DCP-o-matic v1's gamma correction for Rec 601/709.
Dennis points out that `there definitely *is* no right answer for
BT.709 gamma' and `your formula [the DCP-o-matic v2 gamma function
before this commmit] is what I'm calling the inverse of the BT.709
camera transfer function.  The EBU would not have bothered to write:
"Therefore the monitor gamma is not, and never has been, the inverse
of the camera gamma." if no-one ever thought of using it'.

At least the 2.2 gamma of v1 should not surprise anybody.