ClairMeta points out that SMPTE subtitle MXF's ResourceID should be the same as the internal XML ID.  This in turn must be different from the MXF's UUID.
I *think* that we have:

AssetUUID: MXF's UUID, which must be different; written from
the WriterInfo (set up by fill_writer_info, which is passed _id)

AssetID: MXF's ResourceID, which must be the _xml_id; written from
the TimedTextDescriptor.

I believe that this patch fixes the ResourceID to be the XML ID (as
required) without undoing the Doremi fix whereby the MXF UUID must
be different to the XML ID.

This patch quells ClairMeta's error.