Fix failure to verify when the XSD/DTD path has a space in it.
With spaces in the path we ended up with spaces in the string
passed to setExternalSchemaLocation which screws everything up.
It turns out we can remove the URIs from that string, which seems
to move the resolution between file and URI to the LocalFileResolver,
which may now be asked to resolve either a file or a URI.

Maybe I've missed something obvious here.