Fix terrible SoundAsset::equals() implementation.
It would check individual bytes of samples to see if they differed
by more than the threshold.  Not only is this almost useless, but
the default threshold is 256 so with the default settings it would
always say that two assets of the same length (and channels, etc.)
were the same, even if the sample data was different.