Fix strange crash in verify().
This appears to be necessary to stop the first write to
state.subtitle_language from segfaulting.  It looks like the
optional<> is in some uninitialised / indeterminate state.

I could only reproduce this crash when running verify() from DoM,
not from the `dcpverify` tool.  And some simple test programs suggest
that even doing

State state;

should call the constructors of anything in State.