Fix problems when adding KDMs to a VF, before adding the OV.
If we have a KDM which contains keys for all the assets referred
to by a VF (even the ones in the OV) we would previously throw
an exception because we tried to give the KDM to the resolved
assets when it was given to the Reel.

Here we change things so that the addition of KDMs to assets can
be deferred if the asset is not yet resolved.