Change how whitespace in subtitles is handled.
Previously we would discard any whitespace subtitle content we found.
This had the side-effect of making a verification test fail (checking
that at least one subtitle is in each reel; blank subtitles are often
used to avoid this warning).

Here we take any subtitle content, whitespace or not, inside a
<Text> or <Image> node which I think is a little more correct.